requirements.txt
alembic==1.4.2
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
peewee==3.13.3
psycopg2-binary==2.8.5
PyGreSQL==5.1.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.4
Werkzeug==1.0.1
